+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Unrelease
 
+- JSON document no longer accepts value other than object, array, null
 - Fix bug: upper case E in float number ( 12.34E+123 )
 - UTF-8 string support for ruby-1.8.7